+++ highlighters/main/test-data/assignment-before-resword2.zsh
@@ -33,8 +33,8 @@ BUFFER=$'foo=bar ( :; )'
 expected_region_highlight=(
   '1 7 assign' # foo=bar
   '5 7 default' # bar
-  '9 9 unknown-token "issue #461"' # (
+  '9 9 unknown-token' # (
   '11 11 builtin' # :
   '12 12 commandseparator' # ;
-  '14 14 reserved-word' # )
+  '14 14 unknown-token' # )
 )